move from Header to term component
